+++ duckscript_sdk/src/sdk/std/net/mod.rs
@@ -1,3 +1,4 @@
+mod ftp;
 mod hostname;
 mod http_client;
 mod wget;
@@ -14,6 +15,8 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(hostname::create(&package))?;
     commands.set(http_client::create(&package))?;
     commands.set(wget::create(&package)?)?;
+
+    ftp::load(commands, &package)?;
 
     Ok(())
 }