+++ duckscript_sdk/src/sdk/std/noop/mod_test.rs
@@ -1,0 +1,22 @@
+use super::*;
+use crate::test;
+use crate::test::CommandValidation;
+
+#[test]
+fn common_functions() {
+    test::test_common_command_functions(create(""));
+}
+
+#[test]
+fn run_no_args() {
+    test::run_script_and_validate(vec![create("")], "out = noop", CommandValidation::None);
+}
+
+#[test]
+fn run_multiple_args() {
+    test::run_script_and_validate(
+        vec![create("")],
+        "out = noop 1 2 \"3 4\"",
+        CommandValidation::None,
+    );
+}