+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## CHANGELOG
 
+### v0.6.0
+
+* New ftp_put command.
+* New fpt_put_in_memory command.
+* New ftp_get command.
+* New ftp_get_in_memory command.
+* New ftp_list command.
+* New ftp_nlst command.
+* New noop command.
+* \[Breaking Change\] Conditions to support function calls #116
+
 ### v0.5.0 (2020-06-06)
 
 * Fixed end command bug which directed execution to wrong sub command when multiple script contexts are running #110